crowd_sourcing/requirements.txt
awscli==1.18.83
boto3==1.14.6
botocore==1.17.6
colorama==0.4.3
docutils==0.15.2
jmespath==0.10.0
pyasn1==0.4.8
python-dateutil==2.8.1
PyYAML==5.3.1
rsa==3.4.2
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.9
xmltodict==0.12.0
